pyproject.toml
[project]
name = "agentcore-coldstart-bench"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.0",
    "strands-agents>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "requests>=2.32.0",
]
